// src/lib/types.ts
export interface Product {
  id: number;
  title: string;
  price: number;
  category: string;
  description: string;
  image: string;
  rating?: { rate: number; count: number };
  stock: number; // Add explicit stock field
}

export interface CartItem extends Pick<Product, 'id' | 'title' | 'price' | 'image'> {
  quantity: number;
}